# services/core/api/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Timeline API
  version: 0.2.0
  description: |
    个人时间线博客 REST API。OpenAPI 是前后端唯一接口合同。
    写入接口使用同源 HttpOnly session cookie，并要求 X-CSRF-Token；创建、提交、导出和上传终结支持 Idempotency-Key。
servers:
  - url: /api/v1
...
paths:
...
  /admin/media/{id}/upload:
    head:
      tags: [media]
      operationId: getMediaUploadOffset
      security: [SessionCookie: []]
      parameters: [{$ref: '#/components/parameters/MediaID'}]
      responses: {'204': {description: 当前上传偏移, headers: {Upload-Offset: {schema: {type: integer, format: int64}}, Tus-Resumable: {schema: {type: string}}}}}
    patch:
      tags: [media]
      operationId: appendMediaUpload
      security: [SessionCookie: []]
      parameters: [{$ref: '#/components/parameters/MediaID'}, {$ref: '#/components/parameters/CSRF'}, {$ref: '#/components/parameters/Origin'}, {name: Upload-Offset, in: header, required: true, schema: {type: integer, format: int64}}]
      requestBody: {required: true, content: {application/offset+octet-stream: {schema: {type: string, format: binary}}}}
      responses: {'204': {description: 分片已追加, headers: {Upload-Offset: {schema: {type: integer, format: int64}}, Tus-Resumable: {schema: {type: string}}}}, '409': {$ref: '#/components/responses/Conflict'}}
  /admin/media:
    get:
      tags: [media]
      operationId: listMedia
      security: [SessionCookie: []]
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Cursor'}, {name: q, in: query, schema: {type: string}}]
      responses: {'200': {description: 媒体库, content: {application/json: {schema: {$ref: '#/components/schemas/MediaListResponse'}}}}}
  /admin/media/{id}:
    get:
      tags: [media]
      operationId: getMedia
      security: [SessionCookie: []]
      parameters: [{$ref: '#/components/parameters/MediaID'}]
      responses: {'200': {description: 媒体元数据, content: {application/json: {schema: {$ref: '#/components/schemas/Media'}}}}}
    delete:
      tags: [media]
      operationId: deleteMedia
      security: [SessionCookie: []]
      parameters: [{$ref: '#/components/parameters/MediaID'}, {$ref: '#/components/parameters/CSRF'}, {$ref: '#/components/parameters/Origin'}]
      responses: {'204': {description: 已删除}, '409': {$ref: '#/components/responses/Conflict'}}
  /media/{id}/content:
    get:
      tags: [media]
      operationId: getMediaContent
      parameters: [{$ref: '#/components/parameters/MediaID'}, {name: Range, in: header, schema: {type: string}}, {name: If-None-Match, in: header, schema: {type: string}}]
      responses: {'200': {description: 媒体内容, headers: {ETag: {schema: {type: string}}, Accept-Ranges: {schema: {type: string}}}, content: {application/octet-stream: {schema: {type: string, format: binary}}}}, '206': {description: 部分媒体内容}, '404': {$ref: '#/components/responses/NotFound'}}
...
components:
  securitySchemes:
    SessionCookie: {type: apiKey, in: cookie, name: timeline_session, description: HttpOnly SameSite=Lax session cookie; production sets Secure}
  parameters:
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    Cursor: {name: cursor, in: query, schema: {type: string}}
...
    CSRF: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 16}}
    Origin: {name: Origin, in: header, required: false, schema: {type: string, format: uri}}
...
    MediaID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
...
    Conflict: {description: 乐观锁或资源冲突, content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemConflict'}}}}
    NotFound: {description: 资源不存在, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
...
  schemas:
...
    Visibility: {type: string, enum: [public, private]}
...
    Media: {type: object, required: [id, originalName, mimeType, sizeBytes, visibility, status], properties: {id: {type: string, format: uuid}, originalName: {type: string}, mimeType: {type: string}, sizeBytes: {type: integer, format: int64}, visibility: {$ref: '#/components/schemas/Visibility'}, status: {type: string, enum: [uploading, ready, failed, deleting, deleted]}, sha256: {type: string}, createdAt: {type: string, format: date-time}}}
    MediaListResponse: {type: object, required: [media], properties: {media: {type: array, items: {$ref: '#/components/schemas/Media'}}, nextCursor: {type: string}}}
...
    Problem: {type: object, required: [type, title, status, detail], properties: {type: {type: string, format: uri-reference}, title: {type: string}, status: {type: integer}, detail: {type: string}, instance: {type: string}, requestId: {type: string}, fieldErrors: {type: object, additionalProperties: {type: string}}}}
    ProblemConflict:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties: {serverRevision: {type: integer, format: int64}, conflictCopyId: {type: string, format: uuid}}